refactor(slm): extract compose fleet seeder/heartbeat into services/compose_fleet (#9854)
Move _COMPOSE_NODE_SPECS, _SELF_MANAGED_NODE_IDS, _SLM_MGMT_IP,
_compose_nodes_enabled, _probe_tcp, _node_reachable,
_ensure_compose_nodes and _heartbeat_self_managed_nodes verbatim from
main.py into services/compose_fleet.py, exposing seed_compose_nodes()
and start_compose_heartbeat() as the public API wired from the lifespan.
Behavior-preserving: startup order, exception handling and shutdown
cancellation are unchanged.
